CXXCBC-639: Allow to build both static and shared libraries
To avoid issues with generation of cmake helper scripts the following
changes has been made:

* interface libraries that are used to carry compiler options and flags
  are declared as IMPORTED to avoid unnecessary INTERFACE_LINK
  dependencies for the static library

* jsonsl, asio and couchbase_backtrace libraries are not linked as a
  targets, but rather as a object files directly. This is also done to
  avoid announcing them as a transitive dependencies. All these
  libraries are hidden and never exposed.

* for other third-party dependencies do not disable install tasks, which
  generate export files

* the static library repackaged and all dependencies included into
  single library file, because of that some of the object files have to
  be renamed with "public_" or "columnar_" prefixes to avoid collisions.

This patch also removes some unused cmake files (of crypto, sasl and
operations/management directories).

Fixes #701

+++ tools/upsert.cxx
@@ -70,7 +70,7 @@ read_file_content(const std::string& filename) -> couchbase::codec::encoded_valu
   file.seekg(0, std::ios::beg);
 
   couchbase::codec::encoded_value result;
-  result.data.resize(size, std::byte{ 0 });
+  result.data.resize(static_cast<std::size_t>(size), std::byte{ 0 });
 
   if (!file.read(reinterpret_cast<char*>(result.data.data()), size)) {
     throw std::filesystem::filesystem_error(
